+++ spec/fixtures/vcr/WebPurify_ImageFilters/imgcheck/returns_an_image_ID.yml
