+++ spec/fixtures/vcr/WebPurify_TextFilters/check/safe_text_returns_false.yml
