+++ spec/fixtures/vcr/WebPurify_TextFilters/check/text_with_profanity_returns_true.yml
