+++ spec/fixtures/vcr/WebPurify_TextFilters/check_count/no_profanities_returns_0.yml
