+++ spec/fixtures/vcr/WebPurify_TextFilters/check_count/one_profanity_returns_1.yml
