+++ spec/fixtures/vcr/WebPurify_TextFilters/check_count/two_profanities_returns_2.yml
