+++ spec/web_purify/methods/text_filters_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe WebPurify::Filters, "#check", vcr: true do
+describe WebPurify::TextFilters, "#check", vcr: true do
   let(:client) { WebPurify::Client.new(ENV["WEBPURIFY_API_KEY"]) }
   let(:options) { {} }
   subject { client.check(text, options) }
@@ -71,7 +71,7 @@ describe WebPurify::Filters, "#check", vcr: true do
 
 end
 
-describe WebPurify::Filters, "#check_count", vcr: true do
+describe WebPurify::TextFilters, "#check_count", vcr: true do
   let(:client) { WebPurify::Client.new(ENV["WEBPURIFY_API_KEY"]) }
   subject { client.check_count(text) }
 
@@ -101,7 +101,7 @@ describe WebPurify::Filters, "#check_count", vcr: true do
 
 end
 
-describe WebPurify::Filters, "#replace", vcr: true do
+describe WebPurify::TextFilters, "#replace", vcr: true do
   let(:client) { WebPurify::Client.new(ENV["WEBPURIFY_API_KEY"]) }
   subject { client.replace(text, "*") }
 
@@ -123,7 +123,7 @@ describe WebPurify::Filters, "#replace", vcr: true do
 
 end
 
-describe WebPurify::Filters, "#return", vcr: true do
+describe WebPurify::TextFilters, "#return", vcr: true do
   let(:client) { WebPurify::Client.new(ENV["WEBPURIFY_API_KEY"]) }
   subject { client.return(text) }
 
@@ -149,39 +149,6 @@ describe WebPurify::Filters, "#return", vcr: true do
     it "returns an array with two profanities" do
       expect(subject).to contain_exactly("damn", "hell")
     end
-  end
-
-end
-
-describe WebPurify::Filters, "#imgcheck", vcr: true do
-  let(:client) { WebPurify::Client.new(api_key: ENV["WEBPURIFY_API_KEY"], service: :image) }
-  let(:imgurl) { "http://www.example.com/test.jpg" }
-  subject { client.imgcheck(imgurl) }
-
-  it "returns an image ID" do
-    expect(subject).to be_instance_of(String)
-  end
-
-end
-
-describe WebPurify::Filters, "#imgstatus", vcr: true do
-  let(:client) { WebPurify::Client.new(api_key: ENV["WEBPURIFY_API_KEY"], service: :image) }
-  let(:imgurl) { "http://www.example.com/test.jpg" }
-  let(:imgid) { client.imgcheck(imgurl) }
-  subject { client.imgstatus(imgid) }
-
-  it "returns the status of image moderation" do
-    expect(subject).to eq("pending")
-  end
-
-end
-
-describe WebPurify::Filters, "#imgaccount", vcr: true do
-  let(:client) { WebPurify::Client.new(api_key: ENV["WEBPURIFY_API_KEY"], service: :image) }
-  subject { client.imgaccount }
-
-  it "returns the number of image submissions remaining" do
-    expect(subject).to be_kind_of(Integer)
   end
 
 end
